[ISSUE #8206]♻️Extract HA, replication, and transfer boundaries
