+++ docs/plans/architecture-refactor-migration/phase-2-core-boundaries/06-storage-local-compatibility-ledger.md
@@ -1,8 +1,9 @@
-# M06-03/M06-04/M06-05 Local 存储兼容与所有权 Ledger
+# M06-03/M06-04/M06-05/M06-06 Local 存储兼容与所有权 Ledger
 
-本文冻结 PR-M06-05 完成后的 Local 存储 canonical owner、`rocketmq-store` 兼容面、后续迁移边界和删除条件。
+本文冻结 PR-M06-06 完成后的 Local 存储 canonical owner、`rocketmq-store` 兼容面、后续迁移边界和删除条件。
 它关闭 CommitLog append/load/recovery、MappedFile、MappedFileQueue、allocation service、flush/group-commit、ConsumeQueue 与 Index
-owner；不会提前关闭 M06-06～M06-08 的 HA、Timer/POP 或 `LocalFileMessageStore` composition。
+以及 backend-neutral HA/replication/transfer owner；不会提前关闭 M06-07～M06-08 的 Timer/POP 或
+`LocalFileMessageStore` composition，也不会把具体 socket、Remoting/controller DTO 或消息存储副作用搬入 Local。
 
 ## Canonical ownership
 
@@ -18,6 +19,7 @@ owner；不会提前关闭 M06-06～M06-08 的 HA、Timer/POP 或 `LocalFileMess
 | FlushManager root、GroupCommit request/batch/worker、queue flush/commit 与 real-time workers | `rocketmq-store-local::flush::{root,group_commit,queue,worker}` | 旧 `DefaultFlushManager` 保持单字段 `FlushManagerRoot<DefaultFlushManagerAdapter>` facade；Store adapter 只持有 config、TaskGroup、mapped-file I/O、checkpoint、health 与消息状态映射 | M06-08 收敛 Local composition，M06-11 冻结最终 Store facade；旧 public path 按下一 major 兼容窗口处理 |
 | ConsumeQueue 20B/46B records、CQExt、single/batch scan/recovery/truncate/search 与 root/store/dispatch driver | `rocketmq-store-local::consume_queue::{record,single,batch,extension,root}` | Store 保留 mapped-file I/O、CQ factory/table、config/checkpoint、CommitLog timestamp lookup、日志与 `DispatchRequest`/CQType 投影；旧 CQ/CQExt public path 精确 re-export 或窄 adapter | M06-08 收敛 Local composition，M06-11 冻结最终 Store facade；持久记录与旧 public path 按下一 major 兼容窗口处理 |
 | Index 40B header、4B slot、20B entry、file put/query、service lifecycle/query/build 与 dispatch root | `rocketmq-store-local::index::{codec,file,service,dispatch}` | `IndexService` 与 dispatcher 保持单字段 Local root facade；Store 仅保留目录/mmap I/O、RwLock/Arc 文件表、config/checkpoint/error flag、shared Java hasher、等待/flush runtime adapter 与日志；`QueryOffsetResult` 精确 re-export | M06-08 收敛 Local composition，M06-11 冻结最终 Store facade；磁盘格式与旧 public path 按下一 major 兼容窗口处理 |
+| HA wire/flow、replication state/progress/ACK、transfer planner/segment/engine/metrics | `rocketmq-store-local::{ha,transfer}` | Store 旧 HA/transfer path 精确 re-export canonical value/engine；具体 socket、controller Remoting DTO、CommitLog/LocalFileMessageStore append、config/logging 与 ServiceContext/TaskGroup lifecycle 作为窄 adapter 保留 | M06-08 收敛 Local store composition，M06-11 冻结最终 Store facade；旧 public path 按下一 major 兼容窗口处理 |
 
 ## Feature compatibility
 
@@ -27,11 +29,13 @@ owner；不会提前关闭 M06-06～M06-08 的 HA、Timer/POP 或 `LocalFileMess
 
 ## Retained Store-only ports
 
-- M06-05 完成后仅保留 Store 外部 adapter：`MessageStoreConfig` 投影、`TaskGroup`/timer/blocking executor、具体 mapped-file
+- M06-06 完成后仅保留 Store 外部 adapter：`MessageStoreConfig` 投影、`TaskGroup`/timer/blocking executor、具体 mapped-file
   I/O、`StoreCheckpoint`、health recorder 与 `GroupCommitStatus -> PutMessageStatus` 映射；这些 adapter 不拥有 Local flush 算法。
 - ConsumeQueue/Index 仅保留具体 mapped-file/file-table/factory、checkpoint、CommitLog timestamp、shared Java hasher、
   `DispatchRequest`/MessageConst/config、error flag、日志以及受审计的 wait/detached-flush side effects；不得复制 Local codec 或 driver。
-- M06-06：HA、replication、transfer 和 controller/replica side effects。
+- HA/replication/transfer 仅保留具体 socket/network、controller Remoting DTO、CommitLog/`LocalFileMessageStore` append、
+  config/logging、`ServiceManager`/`ServiceContext`/`TaskGroup` 与 controller/replica side effects；不得复制 Local wire、
+  flow-control、replication state/progress/ACK 或 transfer driver。
 - M06-07：Timer、POP、revive、stats 与 Local services。
 - M06-08：`LocalFileMessageStore` lifecycle/query/reput/cleanup composition 和 legacy config envelope。
 
@@ -44,13 +48,13 @@ owner；不会提前关闭 M06-06～M06-08 的 HA、Timer/POP 或 `LocalFileMess
 - Local 不依赖 Store facade、Broker、Remoting、RocksDB 或 TieredStore；Rocks/Tiered 不得创建第二 CommitLog。
 - Store facade 只能保留精确 re-export、type alias、composition、config 投影和外部副作用 adapter；不得重新实现已列为 Local owner 的算法。
 - facade 删除必须等到 M06-11/M09 冻结 consumer 清单，并至少跨越下一 major 兼容窗口；没有下游 canonical-import 证据不得删除。
-- PR-M06-03/04/05 可分别按提交/合并整体回滚，因为旧 public path、feature 入口和磁盘数据未改变；不得通过回滚恢复已被契约禁止的双 owner。
+- PR-M06-03/04/05/06 可分别按提交/合并整体回滚，因为旧 public path、feature 入口、wire bytes 和磁盘数据未改变；不得通过回滚恢复已被契约禁止的双 owner。
 
 ## Closeout evidence
 
-- `rocketmq-store-local` all-feature 单元、integration 与 doctest 全绿；Store all-feature lib 544/544。
-- M06 CQ/Flush/Index owner contract 3/3、Store API contract 8/8、StoreLocal ledger/feature/forbidden-edge impacted contract 3/3；
+- `rocketmq-store-local` all-feature 126 个 unit tests、全部 integration/doctest 全绿；Store all-feature lib 531/531。
+- M06 CQ/Flush/Index owner contract 3/3、M06 HA owner contract 2/2、Store API contract 8/8、StoreLocal ledger/feature/forbidden-edge impacted contract 3/3；
   monolithic suite 两次仅因 304/604 秒工具窗口超时未形成结果，未计为通过，受影响契约已拆分重跑全绿。
 - Local/Store strict Clippy、workspace `--no-deps --all-targets --all-features -D warnings` Clippy 与 workspace fmt 全绿。
-- architecture dependency、ArcMut zero-growth、AGENTS routing、enforcing runtime audit 与 diff check 全绿。
-- `git fetch origin main` 后分支相对 `origin/main` 为 0 behind；PR-M06-05 没有未合入的 main 基线。
+- architecture dependency、ArcMut zero-growth（1,171 identities/3,233 occurrences）、AGENTS routing、enforcing runtime audit 与 diff check 全绿。
+- `git fetch origin main` 后分支相对 `origin/main` 为 0 behind；PR-M06-06 没有未合入的 main 基线。
