+++ rocketmq-store-local/src/ha/transfer_engine/vectored.rs
