+++ rocketmq-store-local/src/lib.rs
@@ -17,6 +17,8 @@ pub mod commit_log;
 pub mod config;
 pub mod consume_queue;
 pub mod flush;
+pub mod ha;
 pub mod index;
 pub mod mapped_file;
+pub mod transfer;
 pub mod utils;
