+++ rocketmq-store-local/src/transfer.rs
@@ -12,4 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use rocketmq_store_local::transfer::segment::*;
+pub mod batch;
+pub mod error;
+pub mod planner;
+pub mod segment;
