+++ rocketmq-store/src/ha/transfer_engine.rs
@@ -12,255 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(unix)]
-use crate::ha::transfer_engine::sendfile::SendfileWriteTarget;
-use ::bytes::Bytes;
-use tokio::io::AsyncWrite;
-
-use crate::ha::transfer_engine::bytes::BytesTransferEngine;
-#[cfg(unix)]
-use crate::ha::transfer_engine::sendfile::SendfileTransferEngine;
-use crate::ha::transfer_engine::vectored::VectoredTransferEngine;
-use crate::transfer::batch::TransferBatch;
-use crate::transfer::error::TransferError;
-use crate::transfer::error::TransferResult;
-
-pub mod bytes;
-pub mod sendfile;
-pub mod vectored;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub enum TransferEngineKind {
-    Bytes,
-    Vectored,
-    Sendfile,
-    IoUring,
+pub mod bytes {
+    pub use rocketmq_store_local::ha::transfer_engine::bytes::*;
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TransferEnginePreference {
-    Auto,
-    Bytes,
-    Vectored,
-    Sendfile,
-    IoUring,
+pub mod sendfile {
+    #[allow(
+        unused_imports,
+        reason = "the legacy sendfile module remains present on non-Unix targets"
+    )]
+    pub use rocketmq_store_local::ha::transfer_engine::sendfile::*;
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct TransferEngineAvailability {
-    pub vectored_write_available: bool,
-    pub sendfile_available: bool,
-    pub io_uring_available: bool,
+pub mod vectored {
+    pub use rocketmq_store_local::ha::transfer_engine::vectored::*;
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TransferEngineSelection {
-    pub engine: TransferEngineKind,
-    pub fallback_reason: Option<&'static str>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct TransferStats {
-    pub engine: TransferEngineKind,
-    pub bytes_written: usize,
-    pub body_bytes: usize,
-    pub frame_count: usize,
-    pub write_call_count: usize,
-    pub sendfile_call_count: usize,
-    pub sendfile_bytes: usize,
-    pub fallback_bytes: usize,
-    pub partial_write_count: usize,
-}
-
-pub fn select_transfer_engine(
-    preference: TransferEnginePreference,
-    vectored_write_available: bool,
-) -> TransferEngineSelection {
-    select_transfer_engine_with_availability(
-        preference,
-        TransferEngineAvailability {
-            vectored_write_available,
-            sendfile_available: false,
-            io_uring_available: false,
-        },
-    )
-}
-
-pub fn select_transfer_engine_with_availability(
-    preference: TransferEnginePreference,
-    availability: TransferEngineAvailability,
-) -> TransferEngineSelection {
-    match preference {
-        TransferEnginePreference::Auto if availability.sendfile_available => TransferEngineSelection {
-            engine: TransferEngineKind::Sendfile,
-            fallback_reason: None,
-        },
-        TransferEnginePreference::Auto => select_vectored_transfer_engine(availability.vectored_write_available),
-        TransferEnginePreference::Bytes => TransferEngineSelection {
-            engine: TransferEngineKind::Bytes,
-            fallback_reason: None,
-        },
-        TransferEnginePreference::Vectored => select_vectored_transfer_engine(availability.vectored_write_available),
-        TransferEnginePreference::Sendfile if availability.sendfile_available => TransferEngineSelection {
-            engine: TransferEngineKind::Sendfile,
-            fallback_reason: None,
-        },
-        TransferEnginePreference::Sendfile if availability.vectored_write_available => TransferEngineSelection {
-            engine: TransferEngineKind::Vectored,
-            fallback_reason: Some("sendfile unavailable"),
-        },
-        TransferEnginePreference::Sendfile => TransferEngineSelection {
-            engine: TransferEngineKind::Bytes,
-            fallback_reason: Some("sendfile and vectored write unavailable"),
-        },
-        TransferEnginePreference::IoUring if availability.io_uring_available => TransferEngineSelection {
-            engine: TransferEngineKind::IoUring,
-            fallback_reason: None,
-        },
-        TransferEnginePreference::IoUring if availability.vectored_write_available => TransferEngineSelection {
-            engine: TransferEngineKind::Vectored,
-            fallback_reason: Some("io_uring unavailable"),
-        },
-        TransferEnginePreference::IoUring => TransferEngineSelection {
-            engine: TransferEngineKind::Bytes,
-            fallback_reason: Some("io_uring and vectored write unavailable"),
-        },
-    }
-}
-
-fn select_vectored_transfer_engine(vectored_write_available: bool) -> TransferEngineSelection {
-    if vectored_write_available {
-        TransferEngineSelection {
-            engine: TransferEngineKind::Vectored,
-            fallback_reason: None,
-        }
-    } else {
-        TransferEngineSelection {
-            engine: TransferEngineKind::Bytes,
-            fallback_reason: Some("vectored write unavailable"),
-        }
-    }
-}
-
-pub enum HaTransferEngine<W> {
-    Bytes(BytesTransferEngine<W>),
-    Vectored(VectoredTransferEngine<W>),
-    #[cfg(unix)]
-    Sendfile(SendfileTransferEngine<W>),
-}
-
-impl<W> HaTransferEngine<W> {
-    pub fn from_selection(writer: W, engine: TransferEngineKind) -> Self {
-        match engine {
-            TransferEngineKind::Bytes => Self::Bytes(BytesTransferEngine::new(writer)),
-            TransferEngineKind::Vectored | TransferEngineKind::IoUring => {
-                Self::Vectored(VectoredTransferEngine::new(writer))
-            }
-            #[cfg(unix)]
-            TransferEngineKind::Sendfile => Self::Sendfile(SendfileTransferEngine::new(writer)),
-            #[cfg(not(unix))]
-            TransferEngineKind::Sendfile => Self::Vectored(VectoredTransferEngine::new(writer)),
-        }
-    }
-
-    pub fn into_inner(self) -> W {
-        match self {
-            Self::Bytes(engine) => engine.into_inner(),
-            Self::Vectored(engine) => engine.into_inner(),
-            #[cfg(unix)]
-            Self::Sendfile(engine) => {
-                let (writer, _) = engine.into_parts();
-                writer
-            }
-        }
-    }
-
-    pub fn kind(&self) -> TransferEngineKind {
-        match self {
-            Self::Bytes(_) => TransferEngineKind::Bytes,
-            Self::Vectored(_) => TransferEngineKind::Vectored,
-            #[cfg(unix)]
-            Self::Sendfile(_) => TransferEngineKind::Sendfile,
-        }
-    }
-}
-
-#[cfg(unix)]
-impl<W> HaTransferEngine<W>
-where
-    W: AsyncWrite + SendfileWriteTarget + Unpin,
-{
-    pub async fn send_batch(&mut self, batch: &TransferBatch) -> TransferResult<TransferStats> {
-        match self {
-            Self::Bytes(engine) => engine.send_batch(batch).await,
-            Self::Vectored(engine) => engine.send_batch(batch).await,
-            Self::Sendfile(engine) => engine.send_batch(batch).await,
-        }
-    }
-}
-
-#[cfg(not(unix))]
-impl<W> HaTransferEngine<W>
-where
-    W: AsyncWrite + Unpin,
-{
-    pub async fn send_batch(&mut self, batch: &TransferBatch) -> TransferResult<TransferStats> {
-        match self {
-            Self::Bytes(engine) => engine.send_batch(batch).await,
-            Self::Vectored(engine) => engine.send_batch(batch).await,
-        }
-    }
-}
-
-pub(crate) fn batch_body_bytes(batch: &TransferBatch) -> TransferResult<Bytes> {
-    let mut chunks = batch_body_chunks(batch)?;
-    match chunks.len() {
-        0 => Ok(Bytes::new()),
-        1 => Ok(chunks.remove(0)),
-        _ => {
-            let mut body = Vec::with_capacity(batch.total_body_len);
-            for chunk in chunks {
-                body.extend_from_slice(&chunk);
-            }
-            Ok(Bytes::from(body))
-        }
-    }
-}
-
-pub(crate) fn batch_body_chunks(batch: &TransferBatch) -> TransferResult<Vec<Bytes>> {
-    let mut remaining = batch.total_body_len;
-    if remaining == 0 {
-        return Ok(Vec::new());
-    }
-
-    let mut chunks = Vec::with_capacity(batch.segments.len());
-    for segment in &batch.segments {
-        let bytes = segment.as_bytes().ok_or(TransferError::UnsupportedSegmentSource(
-            "byte-backed segment required for bytes or vectored HA transfer",
-        ))?;
-        let len = bytes.len().min(remaining);
-        if len > 0 {
-            chunks.push(bytes.slice(..len));
-            remaining -= len;
-        }
-        if remaining == 0 {
-            break;
-        }
-    }
-
-    if remaining != 0 {
-        return Err(TransferError::InvalidInput(format!(
-            "transfer batch body underflow: expected {} bytes, missing {} bytes",
-            batch.total_body_len, remaining
-        )));
-    }
-
-    Ok(chunks)
-}
-
-pub(crate) fn write_zero_error() -> TransferError {
-    TransferError::Io(std::io::Error::new(
-        std::io::ErrorKind::WriteZero,
-        "transfer writer returned zero bytes",
-    ))
-}
+pub use rocketmq_store_local::ha::transfer_engine::*;
