+++ scripts/tests/test_m06_store_local_contract.py
@@ -10568,7 +10568,7 @@ def storage_local_compatibility_ledger_violations(
 ) -> list[str]:
     violations: list[str] = []
     required_ledger_fragments = (
-        "# M06-03/M06-04/M06-05 Local 存储兼容与所有权 Ledger",
+        "# M06-03/M06-04/M06-05/M06-06 Local 存储兼容与所有权 Ledger",
         "## Canonical ownership",
         "`rocketmq-store-local::mapped_file`",
         "`rocketmq-store-local::base::allocate_mapped_file_service`",
@@ -10578,10 +10578,11 @@ def storage_local_compatibility_ledger_violations(
         "`rocketmq-store-local::flush::{root,group_commit,queue,worker}`",
         "`rocketmq-store-local::consume_queue::{record,single,batch,extension,root}`",
         "`rocketmq-store-local::index::{codec,file,service,dispatch}`",
+        "`rocketmq-store-local::{ha,transfer}`",
         "## Feature compatibility",
         "`default = []`",
         "## Retained Store-only ports",
-        "- M06-06：",
+        "HA/replication/transfer 仅保留具体 socket/network",
         "- M06-07：",
         "- M06-08：",
         "## Compatibility and removal rules",
@@ -16883,7 +16884,7 @@ mod tests {
         self.assertEqual([], violations())
         mutations = (
             {"candidate_ledger": ledger.replace("`default = []`", "`default = [fast-load]`", 1)},
-            {"candidate_ledger": ledger.replace("- M06-06：", "- M06-09：", 1)},
+            {"candidate_ledger": ledger.replace("- M06-07：", "- M06-09：", 1)},
             {
                 "candidate_local_manifest": local_manifest.replace(
                     "default = []",
